+++ libraries/Geometry/src/Cube.cpp
@@ -16,20 +16,20 @@ Cube::Cube() noexcept {
 // using the optimized slab intersection method
 // https://tavianator.com/2015/ray_box_nan.html
 std::vector<Intersection> Cube::localIntersect(const utility::Ray &transformedRay) const noexcept {
-  float tx1 = (-1 - transformedRay.origin_.x()) / transformedRay.direction_.x();
-  float tx2 = (1 - transformedRay.origin_.x()) / transformedRay.direction_.x();
+  double tx1 = (-1 - transformedRay.origin_.x()) / transformedRay.direction_.x();
+  double tx2 = (1 - transformedRay.origin_.x()) / transformedRay.direction_.x();
 
-  float tmin = std::min(tx1, tx2);
-  float tmax = std::max(tx1, tx2);
+  double tmin = std::min(tx1, tx2);
+  double tmax = std::max(tx1, tx2);
 
-  float ty1 = (-1 - transformedRay.origin_.y()) / transformedRay.direction_.y();
-  float ty2 = (1 - transformedRay.origin_.y()) / transformedRay.direction_.y();
+  double ty1 = (-1 - transformedRay.origin_.y()) / transformedRay.direction_.y();
+  double ty2 = (1 - transformedRay.origin_.y()) / transformedRay.direction_.y();
 
   tmin = std::max(tmin, std::min(ty1, ty2));
   tmax = std::min(tmax, std::max(ty1, ty2));
 
-  float tz1 = (-1 - transformedRay.origin_.z()) / transformedRay.direction_.z();
-  float tz2 = (1 - transformedRay.origin_.z()) / transformedRay.direction_.z();
+  double tz1 = (-1 - transformedRay.origin_.z()) / transformedRay.direction_.z();
+  double tz2 = (1 - transformedRay.origin_.z()) / transformedRay.direction_.z();
 
   tmin = std::max(tmin, std::min(tz1, tz2));
   tmax = std::min(tmax, std::max(tz1, tz2));
