+++ libraries/Geometry/src/Plane.cpp
@@ -8,16 +8,16 @@ namespace raytracer {
 namespace geometry {
 
 Plane::Plane() noexcept {
-  this->setBoundingBox(utility::AABB(utility::Point(-std::numeric_limits<float>::infinity(), 0, -std::numeric_limits<float>::infinity()),
-                                     utility::Point(std::numeric_limits<float>::infinity(), 0, std::numeric_limits<float>::infinity())));
+  this->setBoundingBox(utility::AABB(utility::Point(-std::numeric_limits<double>::infinity(), 0, -std::numeric_limits<double>::infinity()),
+                                     utility::Point(std::numeric_limits<double>::infinity(), 0, std::numeric_limits<double>::infinity())));
 }
 
 std::vector<Intersection> Plane::localIntersect(const utility::Ray &transformedRay) const noexcept {
-  if (std::abs(transformedRay.direction_.y()) < utility::EPSILON<float>) {
+  if (std::abs(transformedRay.direction_.y()) < utility::EPSILON<double>) {
     return std::vector<Intersection>();
   }
 
-  float t = -transformedRay.origin_.y() / transformedRay.direction_.y();
+  double t = -transformedRay.origin_.y() / transformedRay.direction_.y();
   return std::vector<Intersection>{Intersection(this, t)};
 }
