+++ libraries/Utility/include/Color.hpp
@@ -10,24 +10,24 @@ class Color
 {
 public:
    Color() noexcept;
-   Color(float red, float green, float blue) noexcept;
+   Color(double red, double green, double blue) noexcept;
 
   bool operator==(const Color& rhs) const noexcept;
   Color operator+(const Color& rhs) const noexcept;
   Color operator+=(const Color& rhs) noexcept;
   Color operator-(const Color& rhs) const noexcept;
-  Color operator*(const float& rhs) const noexcept;
+  Color operator*(const double& rhs) const noexcept;
   Color operator*(const Color& rhs) const noexcept; // Hadamarad Product
 
   // Color setters and getters
-  const float& red() const noexcept;
-  void red(const float& val) noexcept;
+  const double& red() const noexcept;
+  void red(const double& val) noexcept;
 
-  const float& green() const noexcept;
-  void green(const float& val) noexcept;
+  const double& green() const noexcept;
+  void green(const double& val) noexcept;
 
-  const float& blue() const noexcept;
-  void blue(const float& val) noexcept;
+  const double& blue() const noexcept;
+  void blue(const double& val) noexcept;
 
   utility::Tuple  _color;
 };
