+++ libraries/Utility/include/Matrix.hpp
@@ -18,40 +18,40 @@ template <uint8_t rows, uint8_t cols>
 class Matrix
 {
 private:
-  std::array<float, rows*cols> elements;
+  std::array<double, rows*cols> elements;
 public:
   template<typename... T>
-  explicit Matrix(T... data) noexcept: elements{static_cast<float>(data)...}{};
-  explicit Matrix(const std::array<float, rows*cols>& data) noexcept : elements(data) {}
+  explicit Matrix(T... data) noexcept: elements{static_cast<double>(data)...}{};
+  explicit Matrix(const std::array<double, rows*cols>& data) noexcept : elements(data) {}
 
   bool operator==(const Matrix& rhs) const noexcept;
 
-  const float& at(const std::size_t& row, const std::size_t& column) const noexcept;
-  float& at(const std::size_t& row, const std::size_t& column) noexcept;
+  const double& at(const std::size_t& row, const std::size_t& column) const noexcept;
+  double& at(const std::size_t& row, const std::size_t& column) noexcept;
 
   static Matrix<rows, cols> identity() noexcept;
 
   Matrix<rows, cols>& operator*=(const Matrix<rows, cols>& rhs) noexcept;
   Matrix<rows, cols> transpose() const noexcept;
-  float determinant() const noexcept;
+  double determinant() const noexcept;
   bool invertible() const noexcept;
 };
 
 template <uint8_t rows, uint8_t cols>
-inline const float& Matrix<rows, cols>::at(const std::size_t& row, const std::size_t& column) const noexcept{
+inline const double& Matrix<rows, cols>::at(const std::size_t& row, const std::size_t& column) const noexcept{
   return elements[row*cols+column];
 }
 
 template <uint8_t rows, uint8_t cols> 
-inline float& Matrix<rows, cols>::at(const std::size_t& row, const std::size_t& column) noexcept{
-  return const_cast<float&>(const_cast<const Matrix*>(this)->at(row, column));
+inline double& Matrix<rows, cols>::at(const std::size_t& row, const std::size_t& column) noexcept{
+  return const_cast<double&>(const_cast<const Matrix*>(this)->at(row, column));
 }
 
 template <uint8_t rows, uint8_t cols> 
 bool Matrix<rows, cols>::operator==(const Matrix& rhs) const noexcept{
   return std::equal(std::execution::unseq, this->elements.cbegin(), this->elements.cend(),
                                            rhs.elements.cbegin()  , rhs.elements.cend(), 
-                    floatNearlyEqual<float>);
+                    floatNearlyEqual<double>);
 }
 
 template <uint8_t rows, uint8_t cols> 
@@ -124,21 +124,21 @@ Matrix<rows, cols> Matrix<rows, cols>::transpose() const noexcept{
 }
 
 template<>
-inline float Matrix<1,1>::determinant() const noexcept{
+inline double Matrix<1,1>::determinant() const noexcept{
   auto determinant = this->at(0,0);
 
   return determinant;
 }
 
 template<>
-inline float Matrix<2,2>::determinant() const noexcept{
+inline double Matrix<2,2>::determinant() const noexcept{
   auto determinant = this->at(0,0)*this->at(1,1) - this->at(0,1)*this->at(1,0);
 
   return determinant;
 }
 
 template<>
-inline float Matrix<3,3>::determinant() const noexcept{
+inline double Matrix<3,3>::determinant() const noexcept{
   return
     + this->at(0,0) * (this->at(1,1) * this->at(2,2) - this->at(2,1) * this->at(1,2))
     - this->at(1,0) * (this->at(0,1) * this->at(2,2) - this->at(2,1) * this->at(0,2))
@@ -146,13 +146,13 @@ inline float Matrix<3,3>::determinant() const noexcept{
 }
 
 template<>
-inline float Matrix<4,4>::determinant() const noexcept{
-  const float SubFactor00 = this->at(2,2) * this->at(3,3) - this->at(3,2) * this->at(2,3);
-  const float SubFactor01 = this->at(2,1) * this->at(3,3) - this->at(3,1) * this->at(2,3);
-  const float SubFactor02 = this->at(2,1) * this->at(3,2) - this->at(3,1) * this->at(2,2);
-  const float SubFactor03 = this->at(2,0) * this->at(3,3) - this->at(3,0) * this->at(2,3);
-  const float SubFactor04 = this->at(2,0) * this->at(3,2) - this->at(3,0) * this->at(2,2);
-  const float SubFactor05 = this->at(2,0) * this->at(3,1) - this->at(3,0) * this->at(2,1);
+inline double Matrix<4,4>::determinant() const noexcept{
+  const double SubFactor00 = this->at(2,2) * this->at(3,3) - this->at(3,2) * this->at(2,3);
+  const double SubFactor01 = this->at(2,1) * this->at(3,3) - this->at(3,1) * this->at(2,3);
+  const double SubFactor02 = this->at(2,1) * this->at(3,2) - this->at(3,1) * this->at(2,2);
+  const double SubFactor03 = this->at(2,0) * this->at(3,3) - this->at(3,0) * this->at(2,3);
+  const double SubFactor04 = this->at(2,0) * this->at(3,2) - this->at(3,0) * this->at(2,2);
+  const double SubFactor05 = this->at(2,0) * this->at(3,1) - this->at(3,0) * this->at(2,1);
 
   const Tuple determinantCoefficients{
     + (this->at(1,1) * SubFactor00 - this->at(1,2) * SubFactor01 + this->at(1,3) * SubFactor02),
@@ -167,8 +167,8 @@ inline float Matrix<4,4>::determinant() const noexcept{
 }
 
 template <uint8_t rows, uint8_t cols>
-float Matrix<rows,cols>::determinant() const noexcept{
-  float determinant{};
+double Matrix<rows,cols>::determinant() const noexcept{
+  double determinant{};
 
   for(std::size_t colIndex{0}; colIndex < cols; colIndex++){
     determinant += this->at(0, colIndex)*cofactor(*this, 0, colIndex);
@@ -198,14 +198,14 @@ inline submatrix(const Matrix<rows, cols>& matrix, const std::size_t& skipRow, c
 }
 
 template<uint8_t rows, uint8_t cols>
-float minor(const Matrix<rows, cols>& matrix, const std::size_t& row, const std::size_t& column) noexcept{
+double minor(const Matrix<rows, cols>& matrix, const std::size_t& row, const std::size_t& column) noexcept{
     auto subMatrix =  submatrix(matrix, row, column);
     return subMatrix.determinant();  
 }
 
 template<uint8_t rows, uint8_t cols>
-inline float cofactor(const Matrix<rows, cols>& matrix, const std::size_t& row, const std::size_t& column) noexcept{
-  float cofactor = (row+column)%2 ? -1*minor(matrix, row, column) : minor(matrix, row, column);
+inline double cofactor(const Matrix<rows, cols>& matrix, const std::size_t& row, const std::size_t& column) noexcept{
+  double cofactor = (row+column)%2 ? -1*minor(matrix, row, column) : minor(matrix, row, column);
 
   return cofactor;
 }
@@ -219,29 +219,29 @@ bool Matrix<rows, cols>::invertible() const noexcept{
 template<uint8_t rows, uint8_t cols>
 typename std::enable_if<(rows==4 && cols==4), Matrix<rows, cols>>::type
 inline inverse(const Matrix<rows,cols>& matrix) noexcept{
-  const float Coef00 = matrix.at(2,2) * matrix.at(3,3) - matrix.at(3,2) * matrix.at(2,3);
-  const float Coef02 = matrix.at(1,2) * matrix.at(3,3) - matrix.at(3,2) * matrix.at(1,3);
-  const float Coef03 = matrix.at(1,2) * matrix.at(2,3) - matrix.at(2,2) * matrix.at(1,3);
+  const double Coef00 = matrix.at(2,2) * matrix.at(3,3) - matrix.at(3,2) * matrix.at(2,3);
+  const double Coef02 = matrix.at(1,2) * matrix.at(3,3) - matrix.at(3,2) * matrix.at(1,3);
+  const double Coef03 = matrix.at(1,2) * matrix.at(2,3) - matrix.at(2,2) * matrix.at(1,3);
 
-  const float Coef04 = matrix.at(2,1) * matrix.at(3,3) - matrix.at(3,1) * matrix.at(2,3);
-  const float Coef06 = matrix.at(1,1) * matrix.at(3,3) - matrix.at(3,1) * matrix.at(1,3);
-  const float Coef07 = matrix.at(1,1) * matrix.at(2,3) - matrix.at(2,1) * matrix.at(1,3);
+  const double Coef04 = matrix.at(2,1) * matrix.at(3,3) - matrix.at(3,1) * matrix.at(2,3);
+  const double Coef06 = matrix.at(1,1) * matrix.at(3,3) - matrix.at(3,1) * matrix.at(1,3);
+  const double Coef07 = matrix.at(1,1) * matrix.at(2,3) - matrix.at(2,1) * matrix.at(1,3);
 
-  const float Coef08 = matrix.at(2,1) * matrix.at(3,2) - matrix.at(3,1) * matrix.at(2,2);
-  const float Coef10 = matrix.at(1,1) * matrix.at(3,2) - matrix.at(3,1) * matrix.at(1,2);
-  const float Coef11 = matrix.at(1,1) * matrix.at(2,2) - matrix.at(2,1) * matrix.at(1,2);
+  const double Coef08 = matrix.at(2,1) * matrix.at(3,2) - matrix.at(3,1) * matrix.at(2,2);
+  const double Coef10 = matrix.at(1,1) * matrix.at(3,2) - matrix.at(3,1) * matrix.at(1,2);
+  const double Coef11 = matrix.at(1,1) * matrix.at(2,2) - matrix.at(2,1) * matrix.at(1,2);
 
-  const float Coef12 = matrix.at(2,0) * matrix.at(3,3) - matrix.at(3,0) * matrix.at(2,3);
-  const float Coef14 = matrix.at(1,0) * matrix.at(3,3) - matrix.at(3,0) * matrix.at(1,3);
-  const float Coef15 = matrix.at(1,0) * matrix.at(2,3) - matrix.at(2,0) * matrix.at(1,3);
+  const double Coef12 = matrix.at(2,0) * matrix.at(3,3) - matrix.at(3,0) * matrix.at(2,3);
+  const double Coef14 = matrix.at(1,0) * matrix.at(3,3) - matrix.at(3,0) * matrix.at(1,3);
+  const double Coef15 = matrix.at(1,0) * matrix.at(2,3) - matrix.at(2,0) * matrix.at(1,3);
 
-  const float Coef16 = matrix.at(2,0) * matrix.at(3,2) - matrix.at(3,0) * matrix.at(2,2);
-  const float Coef18 = matrix.at(1,0) * matrix.at(3,2) - matrix.at(3,0) * matrix.at(1,2);
-  const float Coef19 = matrix.at(1,0) * matrix.at(2,2) - matrix.at(2,0) * matrix.at(1,2);
+  const double Coef16 = matrix.at(2,0) * matrix.at(3,2) - matrix.at(3,0) * matrix.at(2,2);
+  const double Coef18 = matrix.at(1,0) * matrix.at(3,2) - matrix.at(3,0) * matrix.at(1,2);
+  const double Coef19 = matrix.at(1,0) * matrix.at(2,2) - matrix.at(2,0) * matrix.at(1,2);
 
-  const float Coef20 = matrix.at(2,0) * matrix.at(3,1) - matrix.at(3,0) * matrix.at(2,1);
-  const float Coef22 = matrix.at(1,0) * matrix.at(3,1) - matrix.at(3,0) * matrix.at(1,1);
-  const float Coef23 = matrix.at(1,0) * matrix.at(2,1) - matrix.at(2,0) * matrix.at(1,1);
+  const double Coef20 = matrix.at(2,0) * matrix.at(3,1) - matrix.at(3,0) * matrix.at(2,1);
+  const double Coef22 = matrix.at(1,0) * matrix.at(3,1) - matrix.at(3,0) * matrix.at(1,1);
+  const double Coef23 = matrix.at(1,0) * matrix.at(2,1) - matrix.at(2,0) * matrix.at(1,1);
 
   const Tuple Fac0{Coef00, Coef00, Coef02, Coef03};
   const Tuple Fac1{Coef04, Coef04, Coef06, Coef07};
@@ -258,7 +258,7 @@ inline inverse(const Matrix<rows,cols>& matrix) noexcept{
   const Tuple SignA{+1, -1, +1, -1};
   const Tuple SignB{-1, +1, -1, +1};
 
-  const float oneOverDeterminant = static_cast<float>(1)/matrix.determinant();
+  const double oneOverDeterminant = static_cast<double>(1)/matrix.determinant();
 
   const Tuple Inv0{(Vec1 * Fac0 - Vec2 * Fac1 + Vec3 * Fac2) * oneOverDeterminant * SignA};
   const Tuple Inv1{(Vec0 * Fac0 - Vec2 * Fac3 + Vec3 * Fac4) * oneOverDeterminant * SignB};
